fix: resolve JSON serialization error and enhance authentication UI
- Fix float32 JSON serialization error in segmentation endpoint
- Update all API calls to use correct port (5328)
- Enhance Auth component with modern design and better UX
- Add Google OAuth sign-in option
- Improve password visibility toggle
- Add proper error handling and loading states
- Update logout functionality with proper state management
- Fix Supabase client imports across components
- Add proper TypeScript interfaces for better type safety

Technical improvements:
- Convert numpy float32 to Python float for JSON serialization
- Update frontend API endpoints to use correct backend port
- Enhance authentication modal with gradient buttons and icons
- Add proper form validation and user feedback
- Implement proper session management

+++ api/pet_segmentation.py
@@ -174,7 +174,7 @@ class PetSegmentation:
         segmented_image = Image.fromarray(segmented_array.astype(np.uint8))
         
         # Calculate confidence (percentage of pixels classified as pet)
-        confidence = mask_np.mean()
+        confidence = float(mask_np.mean())
         
         processing_time = time.time() - start_time
         
@@ -199,8 +199,8 @@ class PetSegmentation:
         return {
             'maskImage': self.encode_image_to_base64(result['mask']),
             'segmentedImage': self.encode_image_to_base64(result['segmented']),
-            'confidence': result['confidence'],
-            'processingTime': result['processing_time']
+            'confidence': float(result['confidence']),  # Convert numpy float32 to Python float
+            'processingTime': float(result['processing_time'])  # Ensure it's a Python float
         }
 
 # Initialize global segmentation model
